Rename module common to jsonrpc
