+++ packages/iov-tendermint-rpc/src/v0-20/requests.ts
@@ -1,7 +1,7 @@
 import { Encoding } from "@iov/encoding";
 
-import { JsonRpcRequest, jsonRpcWith } from "../common";
 import { Base64, Base64String, HexString, notEmpty } from "../encodings";
+import { JsonRpcRequest, jsonRpcWith } from "../jsonrpc";
 import * as requests from "../requests";
 
 /***** queries *****/
