+++ packages/iov-tendermint-rpc/types/adaptor.d.ts
@@ -1,4 +1,4 @@
-import { JsonRpcEvent, JsonRpcRequest, JsonRpcSuccess } from "./common";
+import { JsonRpcEvent, JsonRpcRequest, JsonRpcSuccess } from "./jsonrpc";
 import * as requests from "./requests";
 import * as responses from "./responses";
 import { RpcClient } from "./rpcclient";
