+++ packages/iov-tendermint-rpc/types/v0-20/hasher.d.ts
@@ -1,3 +1,3 @@
 import { PostableBytes } from "@iov/base-types";
-import { TxHash } from "../common";
+import { TxHash } from "../jsonrpc";
 export declare function hashTx(tx: PostableBytes): TxHash;
